Revert AwsClientUtil rename
