+++ grails-aws-sdk-sqs/grails-app/services/grails/plugin/awssdk/sqs/AmazonSQSService.groovy
@@ -11,7 +11,7 @@ import groovy.transform.Synchronized
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.InitializingBean
 
-import static agorapulse.libs.awssdk.util.AwsClientBuilder.configure
+import static agorapulse.libs.awssdk.util.AwsClientUtil.configure
 
 @Slf4j
 class AmazonSQSService implements InitializingBean  {
