+++ grails-aws-sdk-sts/grails-app/services/grails/plugin/awssdk/sts/AmazonSTSService.groovy
@@ -10,7 +10,7 @@ import groovy.transform.stc.SimpleType
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.InitializingBean
 
-import static agorapulse.libs.awssdk.util.AwsClientBuilder.configure
+import static agorapulse.libs.awssdk.util.AwsClientUtil.configure
 
 @Slf4j
 class AmazonSTSService implements InitializingBean {
